[QNN EP] Add neuron_op_test.cc for neuron activation op tests
Move all neuron activation tests (Relu, Sigmoid, Tanh, Gelu, Elu,
Softplus, HardSwish, HardSigmoid) from simple_op_test.cc into a
dedicated neuron_op_test.cc, following the one-file-per-builder
convention established by selu_test.cc, quickgelu_test.cc, etc.
